pyproject.toml
[tool.poetry]
name = "bueno"
version = "2023.12.0"
description = ""
authors = ["Horia Filimonescu <horia.filimonescu+github@gmail.com>"]
readme = "README.md"
packages = [{ include = "bueno", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
flask-cors = "^4.0.0"
flask-mail = "^0.9.1"
flask-httpauth = "^4.8.0"
flask-migrate = "^4.0.5"
flask-marshmallow = "^0.15.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"
apifairy = "^1.3.0"
requests = "^2.31.0"
marshmallow-sqlalchemy = "^0.29.0"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
